+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.5.27 Release:
+## Update Notes:
+- This is not a V14 update. An update for V14 is still a work in progress.
+## Bug Fixes:
+- Underwater
+
 # 1.5.26 Release:
 ## Update Notes:
 - This is not a V14 update. An update for V14 is still a work in progress.